Define UciCommand and UciResponse enums
These cover all of the possible commands we might receive via stdin and
all of the responses we might want to send back.

Some of the types are loosely defined since the 'proper' types (e.g.
Move) are not yet implemented.

For some commands/responses the types aren't clear so those have been
left as `()` with TODOs.