Fix a bug where moves like d7c8 by a king were considered a castle
`distance()` was implemented incorrectly (at least according to how it
was supposed to be used) and would consider a diagonal move as distance
2.

More importantly though, we need to check that we're moving from the
start square (if not, we already forfeited our right to castle). Doing
this removes the need for the distance check anyway.