Allow debugging with perftree
perftree is a tool which allows easily comparing perft results with
Stockfish to debug move generation bugs.

It requires perft-div results to be in a specific format in order to
work.

It also requires a script and the ability to run perft-div based on a
starting FEN + a set of moves.