Clean up player-dependent square code
A lot of code that has the same behaviour but where the squares involved
depend on the current player were using matches, many of which were
duplicated. Factoring these into methods cleans up a lot of these areas.

This also pulls out a castle_rights closure which de-duplicates the
logic for calculating castle rights after a move based on the player
passed as a parameter.