Implement Squares as a wrapper class for bitboard
Squares is, API-wise, just a set of squares. This makes all code dealing
with sets of squares nicer as they don't need to talk in bitboard terms,
but Squares is implemented using bitboards behind the scenes so the
performance is identical.

Also implements Square in terms of a bitboard. Conceptually, then,
Square is just a set of squares that has been validated to contain only
a single square.

Also add tests for FEN parsing and constructing a Square from an index
which broke during this refactoring. FEN parsing should provide fairly
good coverage of most of this logic.