Implement uciproxy and  uci --remote
uciproxy is a replacement for the current out-of-process behaviour.

Typically, chess programs run engines by running the engine process
directly and communicating via stdin and stdout. When debugging, this
can be an issue as when changes are made to the engine code, the whole
game would need to be restarted.

uciproxy works instead by having the engine run a proxy process which
sends its stdin and stdout over the network. The engine can connect to a
known port and consume/send UCI commands over the network instead. If no
engine is running, the proxy process can queue commands so that when the
engine picks up correctly when restarted.

uci --remote connects to the proxy to run the engine, and does its
communication through a TCP socket.