Run Strategy::go in a new thread
This allows UCI commands to continue to be processed on the main thread
while the engine is searching. Communication between the main and
searcher threads is just two bools:

* `stop` - queried by the searcher to see if the main thread wants it to
  stop
* `stopped` - set by the searcher to notify the main thread that it has
  stopped searching and picked (and sent) a best move

The implementation is abstracted behind `Reporter`, but that's mainly to
allow `Strategy` to be used in a non-UCI context where communicating its
best move needs to be done some other way. It might make sense to split
out the concurrency part from the `Reporter` struct down the line.