Add eval for mate and stalemate
Previously, mate would have been i32::MAX. The issue is that this
propagates up the search so any move that results in mate is considered
equal - as a result, the engine will not prioritise root moves that
result in mate in fewer total moves.

Also adds an explicit check for stalemate, ensuring the game search does
not continue past these states - as a result, the engine no longer plays
moves that result in a draw, when it has better options.