Use undo_move instead of copying the game state
When I last benchmarked this, copying was more efficient, which seemed
like it was down to good memory performance on my computer.

However, as the size of the Game object increases (as it has and will
continue to as more information gets added for better search & eval) the
cost of copying will increase, while the cost of make + undo should not
increase by as much.

This has a pretty huge impact on perft performance:

        Before: perft(6) = 12.2s
         After: perft(6) = 4.8s