Implement simple time control
Send the amount of time remaining through to the search routine, and
construct a new TimeControl which handles allocating time for making the
move.

Search then checks periodically (currently every 10000 nodes searched)
to see if we should stop the search, and returns an Err if so to signal
an incomplete search.

Currently the time allocation per move is very simple - it does not
account for increment, and just returns a static amount of time based on
three possible intervals of remaining time.

This also means the max search depth is no longer hardcoded! We'll
search as deep as we have time for.