Add a transposition table
This allows us to store the result of previous computations against the
Zobrist hash of the position.

In addition to shortcutting our search if we already visited a
particular node at a stronger depth, also save the best move we
could've made in each position even if it doesn't beat alpha. Since we
have this, we can always search it first in move_ordering to try and
make more cutoffs.