Don't special-case the root of negamax
Instead of having a separate function for the root to return best_move,
since we are collecting the pv, we can use what was previously
negamax_inner and just take the best move from the pv directly instead.

This saves a bunch of code duplication, but will be more important as
there are more enhancements to negamax which would've required changes
to both functions.