perf: Avoid lots of move legality checking
We don't yet have a fully legal move generator. As a result, we're using
pseudo_legal move generation and checking legality.

The old implementation eagerly checks legality before starting to look
through moves. But this is really wasteful - if we have a beta cutoff
early, we've wasted time checking the legality of all of the other
moves.

Instead, check only once we're about to make a particular move.
In the future, we can avoid this ugliness by just only generating legal
moves.