Fix a zobrist bug due to tracking the current player
When constructing the initial zobrist hash, we correctly want to only
hash in SIDE_TO_PLAY (previously BLACK_TO_PLAY) when it's black's turn.

However, when toggling the position, we always want to XOR this
contribution (in or out). But when it's white's turn, we XOR 0 (since
side_to_play(White) was 0) and thus don't modify the hash.

This gives us bugs where positions with all of the same pieces in the
same positions which differ only by which player's turn it is receive
the same hash.