Only try collecting up to 'depth' moves when retrieving PV
It doesn't make sense to try and collect infinitely here, we know the
depth we've searched so we can't have a PV longer than that.

Looping infinitely terminates when we run out of exact nodes, but if we
have a cycle in the TT then we do actually loop infinitely!