Add a simple test harness for testing expected best moves
This should help prevent search regressions - when there are known best
moves in certain positions we can assert that we find them. We can also
add a similar kind of test when we've made certain blunders in the past.

Currently only a single expected move is supported, but it likely makes
more sense to test, for most cases, that one of a few good moves is
picked.