Don't panic when finding non-exact nodes in TT while gathering PV
If we stumble onto a non-exact node, the PV is shorter than the depth we
searched to. This is valid, as if the principal variation ends in a draw
or mate, we can search to deeper depths than the number of moves left in
the game but the PV will not grow.