Avoid losing on time in extreme time situations
Time management is still very simple, but in extreme time management
situations we still sometimes lose on time as the clock ticks down by a
minute amount move-by-move.

This is particularly problematic when testing the engine with extremely
short time controls, as the game is often decided by a timeout.