Do not allow exiting the search while at a root node
If we're very tight on time, we may have allocated less time for a move
than it takes to prepare the search (currently made worse because we
reallocate all of the memory for the TT each search).

This can result in us quitting the search when state.nodes_visited = 0
(since state.nodes_visited % 10000 == 0 and time_control.should_stop()
will be true) and as a result will have no PV and thus not report a best
move.

After this change, we will at least search to depth 1, and perhaps more
since we only check whether we should terminate every 10,000 nodes. We
should probably see if we can check more often without affecting
performance.