Remove the 'Squares' abstraction
Squares was a wrapper for Bitboard which exposed an API more akin to a
'collection of squares' than a 'collection of bits'.

However, in practice, although it performs the same (it's essentially
optmised away during compilation) it did result in a lot of duplication,
and shuffling between Squares and Bitboard.