Only compute their_attacks in movegen if the king can castle
We don't need to know their_attacks all the time, only if the king is on
the start square and can still castle, since we only use it to check
whether the squares the king needs to move through are attacked.

So, we remove it from movegen::Ctx and compute it in generate_king_moves
if needed.