Only look at attackers of king square to compute king_in_check
Previously, we used generate_all_attacks which generates attacks for
every piece on the board (and all the squares they attack).

But we only care about pieces attacking the king square, so we can
compute this backwards (e.g. knight attacks from the king square is the
same as knights which can attack the king square) to check this much
more cheaply.

perft(7)
Before:

    time taken: 116.383077167s
    nps: 27460194

After:

    time taken: 92.424194209s
    nps: 34578628

Search to depth 8:
Before: 7594ms
After: 6829ms