Switch to always overwriting in the transposition table
This is definitely not a good solution by any means, but the existing
solution appears to suffer from filling up the table with exact entries
frmo early in the game, which then reduce the effective table size for
useful positions later in the game.

That seems to be corroborated by self-play, where always-overwrite gets
the following results:

    Score of always-overwrite vs master: 32 - 18 - 50  [0.570] 100
    ...      always-overwrite playing White: 18 - 6 - 26  [0.620] 50
    ...      always-overwrite playing Black: 14 - 12 - 24  [0.520] 50
    ...      White vs Black: 30 - 20 - 50  [0.550] 100
    Elo difference: 49.0 +/- 48.4, LOS: 97.6 %, DrawRatio: 50.0 %

This comes at the cost of shortening our PV if one of the PV nodes gets
overwritten during the search.