Treat any repeated position as a draw
This encourages making repetitive moves when losing, but discourages it
when winning.

It also seems to avoid the graph history interaction problem with the
transposition table. This is where a TT value is used but is incorrect
because the real position draws by repetition. We'll exit in any
position that repeats a previous position before looking at the
transposition table at all - so they'll always be correctly scored as
draws.