Store moves compressed in the transposition table
Previously we stored Move, which is wasteful as each square is
represented by its bitboard - i.e. 64 bits when the maximum value is 64
so could fit in 6 bits.

Here we add TTMove which stores the squares as u8s, resulting in a
reduction in size of search transposition entries from 40 bytes to 16
bytes.