Fix the queen's piece square table
This was accidentally set to the same values as the king (!).

As a result, the queen was encouraged to stay on the back rank and
strongly discouraged from developing or being involved in the game.

This explains a lot of odd behaviour, like the queen often moving to B2
as white, blocking in the rook for much of the game.

Unsurprisingly, correcting this has a huge effect.

Score of qpst vs master: 109 - 34 - 46  [0.698] 189
...      qpst playing White: 54 - 19 - 21  [0.686] 94
...      qpst playing Black: 55 - 15 - 25  [0.711] 95
...      White vs Black: 69 - 74 - 46  [0.487] 189
Elo difference: 145.9 +/- 45.8, LOS: 100.0 %, DrawRatio: 24.3 %