Add a 'wait' extension to uci 'go'
This blocks on the search completing, which allows piping 'go' commands
to the engine to test it. For example:

    echo 'go depth 8' | engine

Previously would have exited immediately, because after the go command
was parsed we'd get EOF and quit. Now we can:

    echo 'go depth 8 wait' | engine

Specifically, this should allow much better benchmarking since we can
use tools like hyperfine to automate the process of repeated runs.