Implement incremental move sorting
Rather than sorting the whole move list, we instead use a selection sort
to find the next best move each time we need a new move to look at.

This performs worse when we need to inspect all the moves, but if we
achieve a large number of cutoffs then we will save a lot of work
sorting.

It also opens the door to partial movegen (i.e. generating captures
first, and only generating non-captures if we don't cut off earlier).

Note that node counts change because the ordering has changed - rather
than preserving the rough order of movegen, moves are swapped elsewhere
in the array.

Score of incrsort vs master: 1453 - 1273 - 906  [0.525] 3632
...      incrsort playing White: 775 - 581 - 460  [0.553] 1816
...      incrsort playing Black: 678 - 692 - 446  [0.496] 1816
...      White vs Black: 1467 - 1259 - 906  [0.529] 3632
Elo difference: 17.2 +/- 9.8, LOS: 100.0 %, DrawRatio: 24.9 %