Implement incremental move sorting
Rather than sorting the whole move list, we instead use a selection sort
to find the next best move each time we need a new move to look at.

This performs worse when we need to inspect all the moves, but if we
achieve a large number of cutoffs then we will save a lot of work
sorting.

It also opens the door to partial movegen (i.e. generating captures
first, and only generating non-captures if we don't cut off earlier).