Revert "Disable incremental compilation in release mode"
This reverts commit 73c521cf97b5c4166b64dbe8acef35cf52663fd9.

For some reason, this massively changes the performance of the
.sort_unstable_by call in move ordering.

When compiling without this, around 14.4s is spent sorting. With this,
around 9.9s is spent instead (on a search to depth 10) - around 45%
slower!

For the full search for depth 10, it's anywhere from 7-10% slower.

This seems to explain the drop in Lichess ratings mid-December, which I
initially thought was due to the movegen changes. After this revert,
we're convincingly faster than v1.0.

STC:

Score of incr vs noincr: 652 - 642 - 416  [0.503] 1710
...      incr playing White: 339 - 313 - 203  [0.515] 855
...      incr playing Black: 313 - 329 - 213  [0.491] 855
...      White vs Black: 668 - 626 - 416  [0.512] 1710
Elo difference: 2.0 +/- 14.3, LOS: 60.9 %, DrawRatio: 24.3 %
SPRT: llr -0.044 (-1.5%), lbound -2.94, ubound 2.94

LTC:

Score of incr vs noincr: 335 - 310 - 221  [0.514] 866
...      incr playing White: 176 - 152 - 105  [0.528] 433
...      incr playing Black: 159 - 158 - 116  [0.501] 433
...      White vs Black: 334 - 311 - 221  [0.513] 866
Elo difference: 10.0 +/- 20.0, LOS: 83.8 %, DrawRatio: 25.5 %
SPRT: llr 0.363 (12.3%), lbound -2.94, ubound 2.94