Allow passing UCI commands to run as command line args
This is simpler than passing the commands by piping to stdin, but
crucially works much better for situations like profiling where the
profiler (e.g. cargo instruments) starts the binary and it's impossible
to get stdin to the binary without typing manually.

Now, we can give the profiler args to pass to the binary. It also
removes the need for a time limit since the binary will exit
automatically when it is done with the commands.

Now that this capability exists, I've also removed the 'wait' extension
for 'go' since any situation when we would've used 'go ... wait' can now
use command line args instead. To make this work, we always block on the
search when args are passed via the command line.

Finally, the Justfile commands for profiling have dropped the time limit
and instead take the commands to profile, e.g.:

    just profile "go depth 10"