Bundle midgame and endgame evals into a PhasedEval struct
This is a similar approach to the one that Stockfish takes. It doesn't
result in any change in rating:

Score of phased-struct vs master: 1642 - 1633 - 1337  [0.501] 4612
...      phased-struct playing White: 886 - 753 - 667  [0.529] 2306
...      phased-struct playing Black: 756 - 880 - 670  [0.473] 2306
...      White vs Black: 1766 - 1509 - 1337  [0.528] 4612
Elo difference: 0.7 +/- 8.4, LOS: 56.2 %, DrawRatio: 29.0 %

But it does make eval calculations more straightforward since they don't
need to be constantly done separately for each phase. As more eval terms
are added this should save more and more time.