Add some tests to validate TT size optimisations
These tests keep track of the current memory usage per transposition
table entry, and also the number of entries that can be held with the
default table size of 256mb.

As optimisations are made to transposition table entry size, these tests
should start to fail (and can be updated) to validate that the size of
entries decreases and the number of entries increases.