Make a panic move if we didn't have enough time to search
Previously we tried to make sure that we always search for a particular
amount of time, but in practice this doesn't guarantee that we always
have time to search.

Instead, we have `panic_move` which is a very simple way to generate a
legal move, which we call if we didn't have enough time to find a PV
move during search.

When SPRT testing at STC, a previous test which removed the minimum
search time resulted in numerous engine crashes. The same change here
with the addition of panic moving does not pass standard STC after
~16000 games:

Elo   | 2.63 +- 4.64
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 0.48 (-2.94, 2.94) [0.00, 5.00]
Games | N: 15972 W: 5996 L: 5875 D: 4101
Penta | [920, 1526, 3023, 1547, 970]

But does pass regression STC:

ELO: 1.81 +- 5.55 [-3.74,7.18] (LOS 74.19%)
LLR: 2.94 (-2.94, 2.94]

And more importantly, the engine no longer crashes.