Only report PV to seldepth
It's possible, in some positions, for all positions to be transpositions
and end up being cut, in which case our actual depth searched may be
less than the depth for a round of iterative deepening.

If that's the case, we only want to print the principal variation as far
as we actually searched.