Fix not storing moves that caused beta cutoffs in the TT
Not storing these was a huge oversight, as it means in nodes were we got
a TT hit that didn't cause an immediate cutoff, we lost the information
about which move was best in that position. In a lot of cases, this
means doing a bunch of extra useless work searching worse moves.

Elo   | 49.70 +- 16.19 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
Games | N: 1274 W: 575 L: 394 D: 305
Penta | [47, 106, 225, 137, 122]