Do PVS by searching first move with full window and the rest with zero-window
Previously we switched to zero-window when we raised alpha. However, in
the case of a node where a bunch of the first moves don't raise alpha,
we will search them all with the full alpha/beta window. If a later move
raises alpha, a lot of this was wasted work.

Our move ordering is good enough that assuming the first move searched
is best and searching the rest with a zero window is an improvement.

Elo   | 3.97 +- 5.28 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 9020 W: 3322 L: 3219 D: 2479
Penta | [392, 800, 2053, 843, 422]