Pack Move information into 16 bits and add more information
We can fit the source and destination in 6 bits each (0-63) so we have 4
bits spare to add extra information about the move which can prevent us
from needing to do extra checks. For example, we can store if the move
is a capture to prevent us from having to check the destination square
for pieces during make_move().

A diagram with the bit usage is included in moves.rs.

Elo   | 7.03 +- 4.89 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.95 (-2.94, 2.94) [0.00, 5.00]
Games | N: 9286 W: 2703 L: 2515 D: 4068
Penta | [257, 1050, 1894, 1132, 310]

This also allows us to remove TTMove and store Move directly in the TT.
This removes the need to map back and forward between Move and TTMove
during search.