Do not allow TT cutoffs while in PV nodes
This has been close to passing at LTC regression for weeks:

Elo   | -0.40 +- 3.15 (95%)
SPRT  | 40.0+0.40s Threads=1 Hash=128MB
LLR   | 1.79 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 22460 W: 7351 L: 7377 D: 7732
Penta | [785, 2272, 5139, 2252, 782]

Although it does not gain, the change in Elo since I last tested this is
significant. Every other engine does this, so my hypothesis is that, in
combination with some other popular feature, this ends up gaining. As I
don't want to re-test this down the road, I'll merge it now, take the
hit in STC and hope it works out in the long run.