Use hard and time stops in time control
Previously, we used a set amount of time per move and always aborted the
search after this amount of time.

This is not ideal, because a lot of time we'll start a round of
iterative deepening which we're very unlikely to finish, and lose a lot
of search time doing work which is never used.

The new approach splits into two time limits - one 'hard' limit, as
before, after which the search is fully aborted, and one 'soft' limit,
after which we don't think it's worth starting a new round of iterative
deepening.

Our 'hard' limit is higher than before - so if we think it's worth
starting a new depth, we're more likely to complete it. But our soft
limit is a little smaller - so we're less likely to start rounds in
general.

This also opens the door to adjusting the soft limit based on aspects of
the search - for example, searching longer if the best move keeps
changing as this might indicate we need to spend more time on the
position. If the best move is very stable, however, it might be the only
good move in the position and we can spend less time searching there.

Elo   | 27.49 +- 11.36 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
Games | N: 2216 W: 859 L: 684 D: 673
Penta | [93, 185, 417, 280, 133]

Elo   | 43.22 +- 14.69 (95%)
SPRT  | 40.0+0.40s Threads=1 Hash=128MB
LLR   | 2.95 (-2.94, 2.94) [0.00, 5.00]
Games | N: 1414 W: 561 L: 386 D: 467
Penta | [52, 116, 252, 179, 108]