Allocate the TT before the first search
This seems to have been the cause of the drop in Elo when we started
initialising the TT before each game. The first search is when we call
.resize() but this can take ~70ms, which for STC SPRT tests can mean the
first move is only searched to depth 1 or 2.

Obviously this can be pretty catastrophic in some situations, so after
the 'fix' we started making more mistakes in the first move out of book.

Elo   | 10.45 +- 7.45 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.95 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 4090 W: 1501 L: 1378 D: 1211
Penta | [140, 342, 996, 389, 178]