Fix double allocation of PersistentState
559f24b introduced a potential issue where we allocated a
PersistentState when starting up the engine.

We also reset the PersistentState between games by creating a new
`Arc<Mutex<PersistentState>>`. This meant that after `ucinewgame` was
run, there could very briefly be two `PersistentState` hanging around in
memory.

The problem here is the TT. Although it shouldn't be an issue in most
places, in very memory constrained environments, we might not have
enough room for two TTs to be allocated at the same time. This was
causing the process to be OOM-killed.

The new approach avoids reallocating PersistentState and instead
introduces methods to re-zero everything within it.