Remove git-version for determining UCI version
This was useful when I was doing a lot of testing locally, but almost
all of this complexity is now handled by OpenBench, so I don't have a
lot of different engine binaries sitting around locally.

I've not looked at one of these version strings for a long time, so it's
time to remove this.

We retain a -dev suffix to differentiate between final releases and
development binaries, which can be removed with `--features release`
during compilation.

This also gets rid of the annoying behaviour where the first rebuild
after fully building the binary would still need to recompile the
engine.