Avoid locking up the UCI thread if 'Hash' is changed during a search
This code previously called `.lock()` which blocks waiting for a lock.
If the search is running, that lock is held by the search thread until
the search ends.

This means no other UCI commands can be run until the search ends. This
is an especially big issue if the search was triggered using `go
infinite` (or just `go`) since the whole process needs to be killed.

Instead, we can tell the user that it's not possible to change the TT
size in the middle of a search.