Add tablebase support and follow tablebase lines
We currently only follow tablebase lines if the current position is
found in the tablebase - they are not utilised during search.

This is a relatively modest improvement, but should help us in cases
where games do not finish early due to adjudication and the engine would
not otherwise have been able to convert.

The current solution uses shakmaty-syzygy which means the board state
has to be converted to shakmaty's style. This is done by writing the
state out to FEN and then parsing it, which is horribly inefficent.
Before using tablebases in search, I'll need to switch the library to
something lower level.