Hide the internal implementation of 'Move'
We may want to change the internal implementation of Move to store more
information and avoid redundant work (such as re-determining if a
specific move is a capture, castle, or en-passant).

In order to do this, we might need to switch to computing things like
the source and destination squares of the move at access time. This
change hides the implementation details in order to make this change
easier down the road.