Match moves from MoveList instead of constructing them directly
In the future, we might want to adapt Move to contain more information
about the move in question, such as whether it's a capture or not.

If we do this, places where we currently have all the information
necessary to construct a move (e.g. when we get move information via
UCI) will no longer be able to easily construct Move directly.

In order to avoid that trouble, a better approach is to run movegen and
pick the matching Move object out, ensuring it has all the correct
metadata with essentially zero extra effort.

This did well enough in SPRT (although it did not complete) to prove
there's no critical bugs:

Elo   | 3.56 +- 12.59 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 0.42 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 1270 W: 370 L: 357 D: 543
Penta | [36, 130, 285, 153, 31]