Use NonZeroU16 as internal storage for Move data
The Rust compiler is able to perform an optimisation using 'niche'
values in situations like these - it will store the `None` variant of
`Option<NonZeroU16>` as 0. This means we can avoid having to define some
sort of 'null' move and can use the more idiomatic `Option<Move>`, but
generate the same (or similar) machine code.