Add futility pruning
If we're about to finish searching (depth == 1) we only have one move
left to search. If the move is a capture, we could improve the position
a lot, but with a quiet move, the improvement is likely to be much more
limited.

As a result, if our evaluation of the current position is far enough
below our known best evaluation (alpha), we might as well just skip
searching the move altogether.

Elo   | 10.82 +- 6.41 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
Games | N: 5492 W: 1614 L: 1443 D: 2435
Penta | [153, 610, 1077, 725, 181]