Move evaluation tracing into main eval:: module
This prevents the need for a second implementation of the evaluation
function in the tuner. In order to avoid performance regressions, we use
a const generic, TRACE, which should allow us to eliminate the tracing
code when calling from search.

Elo   | 0.39 +- 2.98 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.96 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 23126 W: 6722 L: 6696 D: 9708
Penta | [642, 2542, 5165, 2576, 638]