Break engine/ and tools/ into separate projects
This mainly involves moving some of the tuner code into tools/tuner. Due
to the need to trace during evaluation, the engine/ crate needs some
tuning code - but future changes should be able to move more of the code
that had to stay in engine/ out into tools/tuner.

This also drops the need for the optional CLI in the engine itself. Deps
like clap and indicatif are no longer optional deps of the engine, but
instead required deps of tools/tuner.