Remove the texel tuning code
This is no longer necessary due to the upcoming NNUE evaluation.

However, there's also an insidious bug where the 'tuner' feature is
always enabled for the 'engine' binary, due to it being a sibling of the
'engine' lib.

As a result, since the change in
bf71788518ab008c1d649dba2511b976a5318a47 to use the fact that that
feature is enabled to switch on tracing, the engine has always been
tracing its eval during normal operation.

This results in a significant slowdown:

With accidental tracing: 4939965 nps
With tracing removed: 6142000 nps

bench: 27104020