nnue: Align network to 64
The legacy branch of bullet outputs network weights without alignment,
so it's not possible to use `align(64)`.

Longer term, it'd be great to switch to the non-legacy bullet branch.
But for now, add `tools/alignnet` which just reads the non-aligned net
and outputs an aligned one. It's excluded from `default-members` because
it doesn't compile if the current network file is aligned, since it
transmutes to `UnalignedNetwork` at compile time.

Elo   | 5.51 +- 4.06 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 3.01 (-2.94, 2.94) [0.00, 5.00]
Games | N: 12794 W: 3761 L: 3558 D: 5475
Penta | [337, 1421, 2717, 1546, 376]

bench: 25373227