Hide implementation details of the transposition table
The current TT implementation exposes the exact struct used to store the
entries to the search implementation. As a result, it's hard to change
implementation details of the TT without knock-on effects on the search
code (for example, if we wanted to pack the entries).

The construction of the TT entries moves into the transposition table.

Additionally, the transposition table implementation was generic over
the inner data (with a view to re-using code for pawn transposition
tables, for example). This is now collapsed for simplicity.

Elo   | 1.50 +- 3.56 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.99 (-2.94, 2.94) [-5.00, 0.00]
Games | N: 14866 W: 4077 L: 4013 D: 6776
Penta | [355, 1599, 3427, 1731, 321]

bench: 19458136